+++ Cargo.toml
@@ -47,7 +47,7 @@ affinidi-did-resolver-cache-sdk = { version = "0.5.4", path = "crates/affinidi-d
 affinidi-did-resolver-cache-server = { version = "0.5.4", path = "crates/affinidi-did-resolver/affinidi-did-resolver-cache-server" }
 did-example = { version = "0.5.3", path = "crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-example" }
 did-peer = { version = "0.6.2", path = "crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-peer" }
-did-webvh = { version = "0.1.3", path = "crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh" }
+did-webvh = { version = "0.1.4", path = "crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh" }
 affinidi-meeting-place = { version = "0.1.10", path = "./crates/affinidi-meeting-place" }
 affinidi-messaging-didcomm = { version = "0.10.8", path = "./crates/affinidi-messaging/affinidi-messaging-didcomm" }
 affinidi-messaging-helpers = { version = "0.10.7", path = "./crates/affinidi-messaging/affinidi-messaging-helpers" }
