+++ crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "did-webvh"
-version = "0.1.3"
+version = "0.1.4"
 description = "Implementation of the did:webvh method in Rust, uses the ssi crate"
 repository.workspace = true
 edition.workspace = true
